add voice of gu

+++ preprocess/line_rerank.py
@@ -22,23 +22,28 @@ preprocess_text_config = config.preprocess_text_config
 # @click.option("--file", default='val', type=str, help="选择排序的文件: 'train' 或 'val'")
 def rerank(file):
     # 根据传入参数选择排序的文件
-    file_list = preprocess_text_config.train_path if file == 'train' else preprocess_text_config.val_path
+    # file_list = preprocess_text_config.train_path if file == 'train' else preprocess_text_config.val_path
     # file_list = file_list.replace('train', 'result').replace('val', 'result')
 
-    # 读取文件内容
-    with open(file_list, "r", encoding="utf-8") as f:
-        lines = f.readlines()
-    # 打印读取到的文件行数
-    print(f"正在重新排序文件：{file_list}，总共{len(lines)}行")
+    train_list = preprocess_text_config.train_path
+    val_list = preprocess_text_config.val_path
+    result_list = val_list.replace('train', 'result').replace('val', 'result')
+    for file_list in [train_list, val_list, result_list]:
+        if os.path.exists(file_list):
+            # 读取文件内容
+            with open(file_list, "r", encoding="utf-8") as f:
+                lines = f.readlines()
+            # 打印读取到的文件行数
+            print(f"正在重新排序文件：{file_list}，总共{len(lines)}行")
 
-    # 按行中的数字进行排序
-    sorted_lines = sorted(lines, key=lambda line: extract_numbers(line))
-    # 将排序后的内容重新写回文件
-    with open(file_list, "w", encoding="utf-8") as f:
-        f.writelines(sorted_lines)
+            # 按行中的数字进行排序
+            sorted_lines = sorted(lines, key=lambda line: extract_numbers(line))
+            # 将排序后的内容重新写回文件
+            with open(file_list, "w", encoding="utf-8") as f:
+                f.writelines(sorted_lines)
 
-    # 打印处理完成信息
-    print(f"重新排序完成，总共{len(sorted_lines)}行")
+            # 打印处理完成信息
+            print(f"重新排序完成，总共{len(sorted_lines)}行")
 
 
 # 提取第一个数字和下划线 _ 后的第二个数字
